Phase 79 (HS-79-02): routes/system.py becomes the system package — done
1,299 lines, one build_system_router holding five unrelated families ->
five single-concern routers composed under the unchanged public
constructor: health 59, coders 340, settings 701, voice 202, ws 63,
plus _shared 49 (the state-shape helpers health and coders both use).
settings.py stays over the P63 module budget as ONE concern (the PUT
validation matrix) — it gets a named budget in HS-79-04, not a silent
split.

Verbatim by the programmatic check (zero non-import body lines differ;
19 lazy in-body imports retargeted one package level deeper). Route
paths byte-identical: the regenerated api-surface diff is module
fields only. Tests unmodified: unit 2407, integration 685.
