+++ holdspeak/web/routes/system/_shared.py
@@ -1,0 +1,49 @@
+"""Shared state-shape helpers for the system routers (moved verbatim, HS-79-02)."""
+from __future__ import annotations
+
+from typing import Any, Optional
+
+def _meeting_active_from_state(state: dict[str, Any]) -> bool:
+    if isinstance(state.get("meeting_active"), bool):
+        return bool(state.get("meeting_active"))
+    return bool(state.get("started_at")) and not bool(state.get("ended_at"))
+
+
+def _meeting_summary_from_state(state: dict[str, Any]) -> Optional[dict[str, Any]]:
+    if not _meeting_active_from_state(state):
+        return None
+    meeting_id = state.get("id")
+    if not meeting_id:
+        return None
+    return {
+        "id": meeting_id,
+        "title": state.get("title"),
+        "tags": state.get("tags") if isinstance(state.get("tags"), list) else [],
+        "started_at": state.get("started_at"),
+        "ended_at": state.get("ended_at"),
+        "duration": state.get("duration"),
+        "formatted_duration": state.get("formatted_duration"),
+    }
+
+
+def _normalize_runtime_status_payload(raw_payload: dict[str, Any], state: dict[str, Any]) -> dict[str, Any]:
+    payload = dict(raw_payload)
+    payload_state = payload.get("state")
+    if not isinstance(payload_state, dict):
+        payload_state = state
+    meeting_active = (
+        bool(payload.get("meeting_active"))
+        if isinstance(payload.get("meeting_active"), bool)
+        else _meeting_active_from_state(payload_state)
+    )
+    payload["status"] = payload.get("status") or "ok"
+    payload["mode"] = payload.get("mode") or "web"
+    payload["meeting_active"] = meeting_active
+    payload["state"] = payload_state
+    if "meeting_id" not in payload:
+        payload["meeting_id"] = payload_state.get("id") if meeting_active else None
+    if "meeting" not in payload:
+        payload["meeting"] = _meeting_summary_from_state(payload_state)
+    return payload
+
+
