+++ holdspeak/web/routes/system/voice.py
@@ -1,0 +1,202 @@
+"""The voice lane: wake type, hub transcribe, the preview one-shots, the command test.
+
+Bodies moved verbatim from routes/system.py (HS-79-02, the Phase-63 discipline).
+"""
+from __future__ import annotations
+
+import json
+import re
+from copy import deepcopy
+from datetime import datetime, timezone
+from typing import Any, Optional
+from urllib.parse import urlparse
+
+from fastapi import APIRouter, Request, WebSocket, WebSocketDisconnect
+from fastapi.responses import JSONResponse
+
+from ....logging_config import get_logger
+from ...context import WebContext
+from ...runtime_support import error_500
+
+log = get_logger("web.routes.system")
+
+
+def build_voice_router(ctx: WebContext) -> APIRouter:
+    router = APIRouter()
+
+    @router.post("/api/dictation/wake/type")
+    async def api_wake_type(payload: dict[str, Any]) -> Any:
+        """HS-60: type a stored wake preview, exactly once.
+
+        The token was minted server-side when the preview was created; the
+        runtime types ONLY its own stored text and burns the token. Client
+        text is never accepted here.
+        """
+        if ctx.on_wake_type is None:
+            return JSONResponse(
+                {"success": False, "error": "Wake typing is unavailable in this runtime."},
+                status_code=503,
+            )
+        token = str((payload or {}).get("token", "")).strip()
+        if not token:
+            return JSONResponse(
+                {"success": False, "error": "A preview token is required."},
+                status_code=400,
+            )
+        typed = ctx.on_wake_type(token)
+        if typed is None:
+            return JSONResponse(
+                {"success": False, "error": "Unknown or already used preview token."},
+                status_code=404,
+            )
+        return {"success": True, "typed": typed}
+
+    @router.post("/api/dictation/transcribe")
+    async def api_transcribe(request: Request) -> Any:
+        """HS-78-01: speak-to-fill — browser-captured audio in, text out.
+
+        Accepts one WAV (16 kHz mono, 16-bit PCM) body and runs the
+        runtime's OWN transcriber (one model, one lock) + the dictation
+        punctuation pass. The audio is never persisted and nothing
+        egresses (local Whisper); the route rides the same
+        loopback/token posture as every other route. Size-capped.
+        """
+        if ctx.on_transcribe is None:
+            return JSONResponse(
+                {"success": False, "error": "Transcription is unavailable in this runtime."},
+                status_code=503,
+            )
+        raw = await request.body()
+        if not raw:
+            return JSONResponse(
+                {"success": False, "error": "An audio body is required."}, status_code=400
+            )
+        if len(raw) > 16_000_000:  # ~8 minutes of 16 kHz mono 16-bit
+            return JSONResponse(
+                {"success": False, "error": "Audio too large (cap: 16 MB)."}, status_code=413
+            )
+        try:
+            import io
+            import wave
+
+            import numpy as np
+
+            with wave.open(io.BytesIO(raw)) as wf:
+                if wf.getnchannels() != 1 or wf.getframerate() != 16000 or wf.getsampwidth() != 2:
+                    return JSONResponse(
+                        {
+                            "success": False,
+                            "error": "Expected WAV: 16 kHz, mono, 16-bit PCM.",
+                        },
+                        status_code=400,
+                    )
+                frames = wf.readframes(wf.getnframes())
+            audio = np.frombuffer(frames, dtype=np.int16).astype(np.float32) / 32768.0
+        except Exception:
+            return JSONResponse(
+                {"success": False, "error": "Not a readable WAV body."}, status_code=400
+            )
+        try:
+            text = ctx.on_transcribe(audio)
+        except Exception as exc:
+            log.error(f"speak-to-fill transcription failed: {exc}")
+            return JSONResponse(
+                {"success": False, "error": "Transcription failed."}, status_code=502
+            )
+        return {"success": True, "text": text}
+
+    @router.post("/api/dictation/preview/type")
+    async def api_preview_type(payload: dict[str, Any]) -> Any:
+        """HS-75-01: type a stored hold-key preview, exactly once.
+
+        The token was minted server-side when the preview armed; the
+        runtime types ONLY its own stored text and burns the token. Client
+        text is never accepted here (the wake/type contract).
+        """
+        if ctx.on_preview_type is None:
+            return JSONResponse(
+                {"success": False, "error": "Preview typing is unavailable in this runtime."},
+                status_code=503,
+            )
+        token = str((payload or {}).get("token", "")).strip()
+        if not token:
+            return JSONResponse(
+                {"success": False, "error": "A preview token is required."},
+                status_code=400,
+            )
+        typed = ctx.on_preview_type(token)
+        if typed is None:
+            return JSONResponse(
+                {"success": False, "error": "Unknown or already used preview token."},
+                status_code=404,
+            )
+        return {"success": True, "typed": typed}
+
+    @router.post("/api/dictation/preview/discard")
+    async def api_preview_discard(payload: dict[str, Any]) -> Any:
+        """HS-75-01: burn a stored preview without typing."""
+        if ctx.on_preview_discard is None:
+            return JSONResponse(
+                {"success": False, "error": "Preview discard is unavailable in this runtime."},
+                status_code=503,
+            )
+        token = str((payload or {}).get("token", "")).strip()
+        if not token:
+            return JSONResponse(
+                {"success": False, "error": "A preview token is required."},
+                status_code=400,
+            )
+        if not ctx.on_preview_discard(token):
+            return JSONResponse(
+                {"success": False, "error": "Unknown or already used preview token."},
+                status_code=404,
+            )
+        return {"success": True}
+
+    @router.post("/api/commands/test")
+    async def api_test_voice_command(payload: dict[str, Any]) -> Any:
+        """HS-52-05: fire one voice command action from the board, to verify it.
+
+        Egress kinds (open_url / launch_app / shell) run on the host through the same
+        bounded connector the dispatcher uses (the browser cannot open a terminal). The
+        `type_text` kind types into whatever app has focus when the keyword is spoken, so
+        there is nothing to run here; it returns a preview instead of firing.
+        """
+        from ....config import VoiceMacroAction, VoiceMacroError
+
+        try:
+            action = VoiceMacroAction(
+                kind=str((payload or {}).get("kind", "")),
+                payload=str((payload or {}).get("payload", "")),
+            )
+        except VoiceMacroError as exc:
+            return JSONResponse({"ok": False, "error": str(exc)}, status_code=400)
+
+        if action.kind == "type_text":
+            return JSONResponse({
+                "ok": True,
+                "tested": False,
+                "preview": action.preview(),
+                "note": "types into the focused app",
+            })
+
+        from ....plugins.actuators import ActuatorProposal
+        from ....plugins.voice_macro_connector import build_voice_macro_connector
+
+        proposal = ActuatorProposal(
+            target="voice_macro",
+            action=action.kind,
+            preview=action.preview(),
+            payload={"kind": action.kind, "payload": action.payload},
+            reversible=False,
+            required_capabilities=(),
+        )
+        try:
+            connector = build_voice_macro_connector(action)
+            result = connector(proposal)
+            return JSONResponse({"ok": True, "tested": True, "result": result})
+        except Exception as exc:  # a failed command is reported inline, not as a 5xx
+            return JSONResponse({"ok": False, "error": f"{type(exc).__name__}: {exc}"})
+
+
+    return router
